New Development: No             JIRA Issue: No       Ready to Release: Yes, well maybe      Interface Changes: Yes What Interface Changed: Changed ssize_t to Long          Test Programs: Unit test somewhere   Put in Release Notes: No              Module(s): casacore            Description: Changed ssize_t to long so it will compile on some linux 32bit machines.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/casacore/alma-evla@8604 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ lattices/lattices/Lattices/LatticeExprNode.cc
@@ -254,7 +254,7 @@ LatticeExprNode::LatticeExprNode (uInt constant)
 #endif
 }
 
-LatticeExprNode::LatticeExprNode (ssize_t constant) 
+LatticeExprNode::LatticeExprNode (Long constant) 
 : donePrepare_p   (False),
   dtype_p         (TpFloat),
   isInvalid_p     (False),
